ym_itSense3dUte: added automatic selection of SimulationProtocol file, and added/renamed some SimulationProtocol files.
SimulationProtocol files are used for density compensation calculation, necessary because of ramp sampling. They are created by simulating the sequence in Idea / POET.
If a SimulationProtocol file name is not explicitly selected in the reconPars, this code will now construct one based on the scanner type, gradient mode, field of view, resolution and bandwidth found in the twix data. If a file with the given file name exists, it will be used for DCF calculation. It is the user's responsibility to make sure that any existing files with this name were actually created using the corresponding settings.