Added 'automatic' calculation of the initial point of each readout (ie number of initial points to be discarded)
The first few points in each 3D UTE readout are unstable, because the digital filter of the ADC needs to stabilize first. This code assumes there is a fixed duration for the ADC to stabilize, which is probably system dependent. For the present usage (ie PrismaFit), it has been empirically found to be approximately 13 us. This code calculates the number of points this duration corresponds to based on the dwell time found in the twix data.